+++ trunk/SIGFDomain/src/com/compsis/sigf/domain/IOrigem.java
@@ -1,7 +1,7 @@
 package com.compsis.sigf.domain;
 
 public interface IOrigem {
-	public Integer getIdOrigem();
+	public Long getIdOrigem();
 	public Integer getTipoOrigem();
 	public String getDescricao();
 	public Integer getPracaOrigem();
@@ -9,7 +9,7 @@ public interface IOrigem {
 	public Integer getPistaOrigem();
 	public Integer getSentidoOrigem();
 	public Integer getNumeroCabine();
-	public Integer getIdOrigemPai();
+	public Long getIdOrigemPai();
 	public Integer getTipoPistaArrecadacao();
 	public Integer getFlagOperacao();
 	public Integer getAtiva();
