+++ trunk/SIGFDomain/src/com/compsis/sigf/domain/Praca.java
@@ -387,9 +387,9 @@ public class Praca extends DomainObject implements Replicable, IOrigem{
     }
 
 	@Override
-	public Integer getIdOrigem() {
+	public Long getIdOrigem() {
 		String idorigem = IDUtils.IdOrigem.getIdOrigem(this);
-		return idorigem!=null?Integer.parseInt(idorigem):null;
+		return idorigem!=null?Long.parseLong(idorigem):null;
 	}
 
 	@Override
@@ -435,9 +435,9 @@ public class Praca extends DomainObject implements Replicable, IOrigem{
 	}
 
 	@Override
-	public Integer getIdOrigemPai() {
+	public Long getIdOrigemPai() {
 		String idorigem = IDUtils.IdOrigem.getIdOrigem(this.getConcessionaria());
-		return idorigem!=null?Integer.parseInt(idorigem):null;
+		return idorigem!=null?Long.parseLong(idorigem):null;
 	}
 
 	@Override
